+++ tests/test_cli.py
@@ -1,5 +1,7 @@
 """Tests of the morecantile CLI"""
 
+import json
+
 import pytest
 from click.testing import CliRunner
 
@@ -19,6 +21,33 @@ def test_cli_shapes():
         result.output
         == '{"bbox": [-105.46875, 39.909736, -104.765625, 40.446947], "geometry": {"coordinates": [[[-105.46875, 39.909736], [-105.46875, 40.446947], [-104.765625, 40.446947], [-104.765625, 39.909736], [-105.46875, 39.909736]]], "type": "Polygon"}, "id": "(106, 193, 9)", "properties": {"grid_crs": "http://www.opengis.net/def/crs/EPSG/0/3857", "grid_name": "WebMercatorQuad", "title": "XYZ tile (106, 193, 9)"}, "type": "Feature"}\n'
     )
+
+    result = runner.invoke(
+        cli, ["shapes", "--precision", "6", "--geographic"], "[106, 193, 9]"
+    )
+    assert result.exit_code == 0
+    assert (
+        result.output
+        == '{"bbox": [-105.46875, 39.909736, -104.765625, 40.446947], "geometry": {"coordinates": [[[-105.46875, 39.909736], [-105.46875, 40.446947], [-104.765625, 40.446947], [-104.765625, 39.909736], [-105.46875, 39.909736]]], "type": "Polygon"}, "id": "(106, 193, 9)", "properties": {"grid_crs": "http://www.opengis.net/def/crs/EPSG/0/3857", "grid_name": "WebMercatorQuad", "title": "XYZ tile (106, 193, 9)"}, "type": "Feature"}\n'
+    )
+
+    # With TMS's CRS
+    with pytest.warns(UserWarning):
+        result = runner.invoke(
+            cli, ["shapes", "--precision", "6", "--projected"], "[106, 193, 9]"
+        )
+        assert result.exit_code == 0
+        feature = json.loads(result.output)
+        assert feature["crs"]
+
+    # geographic CRS (non WGS84)
+    with pytest.warns(UserWarning):
+        result = runner.invoke(
+            cli, ["shapes", "--precision", "6", "--crs", "epsg:4150"], "[106, 193, 9]"
+        )
+        assert result.exit_code == 0
+        feature = json.loads(result.output)
+        assert feature["crs"]
 
     # tile as arg
     result = runner.invoke(cli, ["shapes", "[106, 193, 9]", "--precision", "6"])